refactor: extract agent session sync from App
